refactor backend file structure, add typescript config file
