+++ Dockerfile
@@ -17,5 +17,7 @@ RUN apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install --no-instal
 ENV PORT=80 UPDATE_DATA=false
 EXPOSE 80
 
+RUN node_modules/typescript/bin/tsc
+
 ENTRYPOINT ["/usr/bin/node"]
-CMD ["./app.js"]
+CMD ["./backend/app.js"]
