+++ package.json
@@ -2,7 +2,7 @@
   "name": "stellar-dashboard",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "engines": {
     "node": "16.x"
   },
@@ -11,7 +11,8 @@
     "build": "gulp build",
     "heroku-postbuild": "gulp build",
     "lint:prettier": "prettier --config --write '**/*.js'",
-    "test": "DEV=true mocha './test/**/*.js' --exit"
+    "test": "DEV=true mocha './test/**/*.js' --exit",
+    "dev-backend": "UPDATE=true DEV=true ts-node backend/app.ts"
   },
   "husky": {
     "hooks": {
@@ -65,9 +66,15 @@
     "sequelize": "^5.3.0",
     "stellar-sdk": "^2.0.0-beta.7",
     "style-loader": "^0.19.0",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4",
     "webpack": "^1.13.3"
   },
   "devDependencies": {
+    "@stellar/tsconfig": "^1.0.2",
+    "@types/express": "^4.17.13",
+    "@types/express-http-proxy": "^1.6.3",
+    "@types/morgan": "^1.9.3",
     "chai": "^4.3.4",
     "husky": "1.3.1",
     "lint-staged": "7.3.0",
